Move syscall handler private helpers
